% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_function.R
\name{top_starwars}
\alias{top_starwars}
\title{Ranking Variables}
\usage{
top_starwars(variable = "height", top = 10)
}
\arguments{
\item{variable:}{It gets the variable, a numeric column the data}

\item{top:}{A range of ranking}
}
\value{
Return a dataframe of names and the top n entries
}
\description{
Returns a data frame ranked by a user-specified variable. Useful for sorting characters by attributes like height or mass.
}
\examples{
top_starwars("mass", 5);
top_starwars("height", 5);
}
